% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cramer}
\alias{cramer}
\title{Cramer's V statistic}
\usage{
cramer(dfm, x, y)
}
\arguments{
\item{dfm}{data.frame. probability distributions.}

\item{x}{character. name of categorical or discrete variable.}

\item{y}{character. name of another categorical or discrete variable.}
}
\value{
data.frame. It has the following variables.:
\itemize{
\item var1 : character. first variable name.
\item var2 : character. second variable name.
\item chisq : numeric. Chisquare statistic.
\item df : integer. degree of freedom.
\item pval : numeric. p value of Chisquare test.
\item coef_corr : numeric. Cramer's V statistic.
}
}
\description{
Computes the Cramer's V statistic and Chisquare p value between two categorical variables in data.frame.
}
\examples{
cramer(mtcars, "gear", "carb")

}
\seealso{
\code{\link{theil}}.
}
