% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R, R/dlookr-deprecated.R
\name{plot_correlate}
\alias{plot_correlate}
\alias{dlookr-deprecated}
\title{Deprecated functions in package \sQuote{dlookr}}
\usage{
plot_correlate(.data, ...)
}
\description{
These functions are provided for compatibility with older versions
of \sQuote{dlookr} only, and will be defunct at the next release.
}
\details{
The following functions are deprecated and will be made defunct; use 
the replacement indicated below:
\itemize{
 \item{plot_correlate: \code{\link{plot.correlate}}}
}
}
\keyword{internal}
