% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlookr.R
\docType{package}
\name{dlookr-package}
\alias{dlookr}
\alias{dlookr-package}
\title{dlookr: Tools for Data Diagnosis, Exploration, Transformation}
\description{
dlookr provides data diagnosis, data exploration and transformation of variables
during data analysis.
}
\details{
It has three main goals:

\itemize{
\item When data is acquired, it is possible to judge whether data is erroneous
or to select a variable to be corrected or removed through data diagnosis.
\item Understand the distribution of data in the EDA process. It can also
understand the relationship between target variables and predictor variables
for the prediction model.
\item Imputes missing value and outlier to standardization and resolving skewness.
And, To convert a continuous variable to a categorical variable, bin the continuous variables.
}

To learn more about dlookr, start with the vignettes:
`browseVignettes(package = "dlookr")`
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/choonghyunryu/dlookr/}
  \item \url{https://choonghyunryu.github.io/dlookr/}
  \item Report bugs at \url{https://github.com/choonghyunryu/dlookr/issues}
}

}
\author{
\strong{Maintainer}: Choonghyun Ryu \email{choonghyun.ryu@gmail.com}

}
