% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlookr_templ_paged.R
\name{dlookr_orange_paged}
\alias{dlookr_orange_paged}
\alias{dlookr_blue_paged}
\title{Generate paged HTML document}
\usage{
dlookr_orange_paged(...)

dlookr_blue_paged(...)
}
\arguments{
\item{...}{arguments to be passed to \code{pagedown::\link[pagedown]{html_paged}}.}
}
\value{
document of markdown format.
}
\description{
Generate paged HTML document
}
\references{
\url{https://pagedown.rbind.io}
}
