% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlookr_templ_html.R
\name{dlookr_templ_html}
\alias{dlookr_templ_html}
\title{dlookr HTML template

Loads additional style and template file}
\usage{
dlookr_templ_html(toc = TRUE, ...)
}
\arguments{
\item{toc}{should a table of contents be displayed?}

\item{...}{additional arguments provided to \code{html_document}}
}
\value{
An R Markdown output format.
}
\description{
dlookr HTML template

Loads additional style and template file
}
\references{
https://raw.githubusercontent.com/dr-harper/example-rmd-templates/master/R/my_html_format.R
}
