% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_os}
\alias{get_os}
\title{Finding Users Machine's OS}
\usage{
get_os()
}
\value{
OS names. "windows" or "osx" or "linux"
}
\description{
Get the operating system that users machines.
}
\examples{
get_os()

}
