% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{import_liberation}
\alias{import_liberation}
\title{Import Liberation Sans Narrow fonts}
\usage{
import_liberation()
}
\description{
Import Liberation Sans Narrow font to be used when drawing charts.
}
\details{
The Liberation(tm) Fonts is a font family which aims at metric compatibility 
with Arial, Times New Roman, and Courier New. It is sponsored by Red Hat.
Import fonts is older versions of the Liberation(tm) fonts. This is released 
as open source under the GNU General Public License version 2 with exceptions. 
https://fedoraproject.org/wiki/Licensing/LiberationFontLicense
}
