% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{ivtable}
\alias{ivtable}
\title{Calculate Information Value}
\usage{
ivtable(.data, y, x, p = 0.05)
}
\arguments{
\item{.data}{a data.frame or a \code{\link{tbl_df}} or a \code{\link{grouped_df}}.}

\item{y}{character or name. the name of target variable in .data.}

\item{x}{character or name. the name of indicator variable in .data.}

\item{p}{numeric. ratio of records per bin. Default is 0.05. This parameter only accepts values that range is (0, 0.5].}
}
\value{
list. information value table and information value.
Attributes of list is as follows.
\itemize{
\item ivtable : data.frame. information value table. 
  \itemize{
    \item Cutpoint : character. class by binned using cut points.
    \item CntRec : numeric. frequency.
    \item CntGood : numeric. frequency with positive(good) cases.
    \item CntBad : numeric. frequency with negative(bad) cases.
    \item CntCumRec : numeric. cumulate frequency.  
    \item CntCumGood : numeric. cumulate frequency with positive(good) cases.
    \item CntCumBad : numeric. cumulate frequency with negative(bad) cases.
    \item PctRec : numeric. relative frequency. 
    \item GoodRate : numeric. relative frequency with positive(good) cases.
    \item BadRate : numeric. relative frequency with negative(bad) cases.
    \item Odds : numeric. odds.
    \item LnOdds : numeric. ln(odds)
    \item WoE : numeric. weight of evidence.
    \item IV : numeric. information value.
  }
\item iv : numeric. information value.
}
}
\description{
The ivtable() calculate a information value and information value.
}
\details{
When developing the scorecard, use WoE(weight of evidence) and IV(information value). 
This information is useful when binning continuous variables.
}
\examples{
# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# Prepare the target variable (positive : 1, negative : 0)
carseats$US2 <- as.integer(carseats$US) - 1

ivtable(carseats, y = "US2", x = "Advertising")
ivtable(carseats, y = US2, x = Advertising)

}
\seealso{
\code{\link{binning_by}}, \code{\link{print.bins}}, \code{\link{summary.bins}}.
}
