% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobchange.R
\docType{data}
\name{jobchange}
\alias{jobchange}
\title{Job Change of Data Scientists}
\format{
A data frame with 19158 rows and 14 variables. The variables are as follows:
\describe{
  \item{enrollee_id}{unique ID for candidate}
  \item{city}{city code.}
  \item{city_dev_index}{developement index of the city (scaled).}
  \item{gender}{gender of candidate.}
  \item{relevent_experience}{relevant experience of candidate}
  \item{enrolled_university}{type of University course enrolled if any.}
  \item{education_level}{education level of candidate.}
  \item{major_discipline}{education major discipline of candidate.}
  \item{experience}{candidate total experience in years.}
  \item{company_size}{number of employees in current employer's company.}
  \item{company_type}{type of current employer.}
  \item{last_new_job}{difference in years between previous job and current job.}
  \item{training_hours}{training hours completed.}
  \item{job_chnge}{if looking for a job change (boolean), Yes, No.}
}
}
\source{
"HR Analytics: Job Change of Data Scientists" in Kaggle <https://www.kaggle.com/arashnic/hr-analytics-job-change-of-data-scientists>, License : CC0(Public Domain
}
\usage{
data(jobchange)
}
\description{
A dataset containing the gender and other attributes of almost 20000 cases.
}
\details{
This dataset designed to understand the factors that lead a person to leave current job for HR researches too.
}
\keyword{datasets}
