% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis of the data}
\usage{
kurtosis(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{logical. Determine whether to remove missing values and calculate them. 
The default is TRUE.}
}
\value{
numeric. calculated kurtosis
}
\description{
This function calculated kurtosis of given data.
}
\examples{
set.seed(123)
kurtosis(rnorm(100))
}
\seealso{
\code{\link{skewness}}.
}
