% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{plot.bins}
\alias{plot.bins}
\title{Visualize Distribution for a "bins" object}
\usage{
\method{plot}{bins}(x, ...)
}
\arguments{
\item{x}{an object of class "bins", usually, a result of a call to binning().}

\item{...}{arguments to be passed to methods, such as graphical parameters (see par).}
}
\description{
Visualize two plots on a single screen.
The plot at the top is a histogram representing the frequency of the level.
The plot at the bottom is a bar chart representing the frequency of the level.
}
\examples{
# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# Binning the carat variable. default type argument is "quantile"
bin <- binning(carseats$Income, nbins = 5)
plot(bin)
# Using another type argument
bin <- binning(carseats$Income, nbins = 5, type = "equal")
plot(bin)
bin <- binning(carseats$Income, nbins = 5, type = "pretty")
plot(bin)
bin <- binning(carseats$Income, nbins = 5, type = "kmeans")
plot(bin)
bin <- binning(carseats$Income, nbins = 5, type = "bclust")
plot(bin)
}
\seealso{
\code{\link{binning}}, \code{\link{print.bins}}, \code{\link{summary.bins}}.
}
