% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{plot.optimal_bins}
\alias{plot.optimal_bins}
\title{Visualize Distribution for an "optimal_bins" Object}
\usage{
\method{plot}{optimal_bins}(x, type = c("all", "dist", "freq", "badrate", "WoE"))
}
\arguments{
\item{x}{an object of class "optimal_bins", usually, a result of a call to binning_by().}

\item{type}{character. options for visualization. Distribution ("dist"), Relateive Frequency ("freq"),
Bad Rate ("badrate"), and Weight of Evidence ("WoE"). and default "all" draw all plot.}
}
\description{
It generates plots for understand distribution, frequency, bad rate, and weight of evidence using optimal_bins.

See vignette("transformation") for an introduction to these concepts.
}
\examples{
# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# optimal binning
bin <- binning_by(carseats, "US", "Advertising")
bin

# summary optimal_bins class
summary(bin)

# information value
attr(bin, "iv")

# information value table
attr(bin, "ivtable")

# visualize all information for optimal_bins class
plot(bin)

# visualize WoE information for optimal_bins class
plot(bin, type = "WoE")

}
\seealso{
\code{\link{binning_by}}, \code{\link{plot.bins}}.
}
