% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{plot.overview}
\alias{plot.overview}
\title{Visualize Information for an "overview" Object}
\usage{
\method{plot}{overview}(x, sort_type = c("none", "name", "type"), ...)
}
\arguments{
\item{x}{an object of class "overview", usually, a result of a call to overview().}

\item{...}{arguments to be passed to methods, such as graphical parameters (see par).}
}
\description{
Visualize a plot by attribute of `overview` class.
Visualize the data type, number of observations, and number of missing values for each variable.
}
\examples{
\donttest{
# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

ov <- overview(carseats)
ov

summary(ov)

plot(ov)

# sort by name of variables
plot(ov, sort_type = "name")

# sort by data type of variables
plot(ov, sort_type = "type")
}

}
\seealso{
\code{\link{overview}}, \code{\link{summary.overview}}.
}
