% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{plot_bar_category}
\alias{plot_bar_category}
\alias{plot_bar_category.data.frame}
\title{Plot bar chart  of categorical variables}
\usage{
plot_bar_category(.data, ...)

\method{plot_bar_category}{data.frame}(.data, ..., top = 10, add_character = TRUE)
}
\arguments{
\item{.data}{a data.frame or a \code{\link{tbl_df}}.}

\item{...}{one or more unquoted expressions separated by commas.
You can treat variable names like they are positions.
Positive values select variables; negative values to drop variables.
If the first expression is negative, plot_bar_category() will automatically
start with all variables.
These arguments are automatically quoted and evaluated in a context where
column names represent column positions.
They support unquoting and splicing.}

\item{top}{an integer. Specifies the upper top rank to extract.
Default is 10.}

\item{add_character}{logical. Decide whether to include text variables in the
diagnosis of categorical data. The default value is TRUE, which also includes character variables.}
}
\description{
The plot_bar_category() to visualizes the distribution of 
categorical data by level or relationship to specific numerical data by level.
}
\details{
The distribution of categorical variables can be understood by 
comparing the frequency of each level. The frequency table helps with this. 
As a visualization method, a bar graph can help you understand 
the distribution of categorical data more easily than a frequency table.
}
\examples{
# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA
set.seed(123)
carseats$Test <- sample(LETTERS[1:15], 400, replace = TRUE)
carseats$Test[1:30] <- NA

# Visualization of all numerical variables
plot_bar_category(carseats)

# Select the variable to diagnose
plot_bar_category(carseats, "ShelveLoc", "Urban")
plot_bar_category(carseats, -ShelveLoc, -Urban)

# Using pipes ---------------------------------
library(dplyr)

# Plot of all categorical variables
carseats \%>\%
  plot_bar_category()

# Visualize just 7 levels of top frequency
carseats \%>\%
  plot_bar_category(top = 7)
   
# Visualize only factor, not character
carseats \%>\%
  plot_bar_category(add_character = FALSE) 
  
}
