% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{plot_box_numeric}
\alias{plot_box_numeric}
\alias{plot_box_numeric.data.frame}
\alias{plot_box_numeric.grouped_df}
\title{Plot Q-Q plot  of numerical variables}
\usage{
plot_box_numeric(.data, ...)

\method{plot_box_numeric}{data.frame}(
  .data,
  ...,
  title = "Box plot by numerical variables",
  each = FALSE
)

\method{plot_box_numeric}{grouped_df}(
  .data,
  ...,
  title = "Box plot by numerical variables",
  each = FALSE
)
}
\arguments{
\item{.data}{data.frame or a \code{\link{tbl_df}} or a \code{\link{grouped_df}}.}

\item{\dots}{one or more unquoted expressions separated by commas.
You can treat variable names like they are positions.
Positive values select variables; negative values to drop variables.
If the first expression is negative, plot_box_numeric() will automatically
start with all variables.
These arguments are automatically quoted and evaluated in a context where
column names represent column positions.
They support unquoting and splicing.}

\item{title}{character. a main title for the plot.}

\item{each}{logical. Specifies whether to draw multiple plots on one screen. 
The default is FALSE, which draws multiple plots on one screen.}
}
\description{
The plot_box_numeric() to visualizes the box plot of numeric data or 
relationship to specific categorical data.
}
\details{
The The box plot helps determine whether the distribution of a numeric variable. 
plot_box_numeric() shows box plots of several numeric variables 
on one screen. This function can also display a box plot for each level of a specific 
categorical variable.
}
\examples{
# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# Visualization of all numerical variables
plot_box_numeric(carseats)

# Select the variable to diagnose
plot_box_numeric(carseats, "Sales", "Income")
plot_box_numeric(carseats, -Sales, -Income)

# Using pipes ---------------------------------
library(dplyr)

# Plot of all numerical variables
carseats \%>\%
  plot_box_numeric()
  
# Using groupd_df  ------------------------------
carseats \%>\% 
  group_by(ShelveLoc) \%>\% 
  plot_box_numeric()
  
carseats \%>\% 
  group_by(ShelveLoc) \%>\% 
  plot_box_numeric(each = TRUE)  
  
}
