% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skewness}
\alias{skewness}
\title{Skewness of the data}
\usage{
skewness(x, na.rm = TRUE)
}
\arguments{
\item{x}{a numeric vector.}

\item{na.rm}{logical. Determine whether to remove missing values and calculate them. 
The default is TRUE.}
}
\value{
numeric. calculated skewness.
}
\description{
This function calculated skewness of given data.
}
\examples{
set.seed(123)
skewness(rnorm(100))
}
\seealso{
\code{\link{kurtosis}}, \code{\link{find_skewness}}.
}
