% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pps.R
\name{summary.pps}
\alias{summary.pps}
\title{Summarizing Predictive Power Score}
\usage{
\method{summary}{pps}(object, ...)
}
\arguments{
\item{object}{an object of class "pps", usually, a result of a call to pps().}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print and summary method for "pps" class.
}
\details{
summary.pps compares the PPS by variables.
}
\examples{
\donttest{
library(dplyr)

# If the target variable is a categorical variable
# Using dplyr
pps_cat <- heartfailure \%>\% 
  target_by(death_event) \%>\% 
  pps()

pps_cat

# summary pps class 
tab <- summary(pps_cat)
tab

##---------------------------------------------------
# If the target variable is a numerical variable
num <- target_by(heartfailure, creatinine)

pps_num <- pps(num)
pps_num

# summary pps class 
summary(pps_num)
}

}
\seealso{
\code{\link{pps}}, \code{\link{plot.pps}}.
}
