% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{theil}
\alias{theil}
\title{Theil's U statistic}
\usage{
theil(dfm, x, y)
}
\arguments{
\item{dfm}{data.frame. probability distributions.}

\item{x}{character. name of categorical or discrete variable.}

\item{y}{character. name of another categorical or discrete variable.}
}
\value{
data.frame. It has the following variables.:
\itemize{
\item var1 : character. first variable name.
\item var2 : character. second variable name.
\item coef_corr : numeric. Theil's U statistic.
}
}
\description{
Computes the Theil's U statistic between two categorical variables in data.frame.
}
\examples{
theil(mtcars, "gear", "carb")

}
\seealso{
\code{\link{cramer}}.
}
