% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{summary.overview}
\alias{summary.overview}
\title{Summarizing overview information}
\usage{
\method{summary}{overview}(object, ...)
}
\arguments{
\item{object}{an object of class "overview", usually, a result of a call to overview().}

\item{...}{further arguments passed to or from other methods.}
}
\description{
print and summary method for "overview" class.
}
\details{
summary.overview() tries to be smart about formatting 14 information of overview.
}
\examples{
\donttest{
ov <- overview(jobchange)
ov

summary(ov)
}

}
\seealso{
\code{\link{overview}}, \code{\link{plot.overview}}.
}
